Extract shared Slack error taxonomy into @vellumai/slack-text
The Slack error-code -> category map, its `SlackErrorCategory` type, and
`classifySlackError` were duplicated verbatim in the assistant
(`messaging/providers/slack/api.ts`) and the gateway (`slack/errors.ts`).
Duplicated logic drifts (root AGENTS.md "Single Source of Truth");
flagged in the #36291 review.

Extract the shared classification into `@vellumai/slack-text` — a package
both already depend on — under a new `./errors` subpath:

- packages/slack-text/src/errors.ts: `SlackErrorCategory`, the code map,
  and `classifySlackError`. The map now includes `msg_blocks_too_long`
  (cumulative Block Kit text over Slack's ~13k ceiling) as a
  `client_error`.
- assistant `api.ts` and gateway `errors.ts` import the classifier from
  the package instead of each keeping a copy. The gateway re-exports
  `classifySlackError` so its existing importers are unaffected, and keeps
  its own gateway-specific `isRetryable` / `getUserMessage` policy (those
  are not duplicated, so they stay where they are used).

The only behavior change is classifying `msg_blocks_too_long`, which fixes
the gateway's `isRetryable` previously treating it as retryable and gives
it a user-facing message. Part of LUM-2618.
